build: use -o pipefail in license_extract.sh
The portable mkfifo approach works fine locally and worked fine in
initial GHA testing, but now it started to run into various stalling
issues on GHA. Presumably depending on whether the pipe is first written
to or read from and attempts to find a simple workaround for it failed.
Sometimes (but not always!) this is accompanied by an error like this:
  cannot open ./__LICENSE_EXTRACT_QUEUE.tmp: Interrupted system call

So convert the script to bash and just use set -o pipefail.

Cherry-picked from: https://github.com/libass/JavascriptSubtitlesOctopus/commit/94cfe111f8f1c672340277b6408e5a12c479469f